Centralize the speeds and fix the ball initial direction and speed
Pull the ball speed and the paddle speed up to the Pong level and
let the Pong controller manage the speed.  This keeps the paddle
and the ball at the same speed.

No longer increase the speed per goal.  Just by time.

Temporarily add a speed debug output so I can watch it.

When the ball is dropped don't allow the sharper angles (generate
each coordinate in the range -1..-0.5 / 0.5..1 getting rid of
all of -0.5..0.5) and then normalize it to the speed length.
This also ensures that the ball speed is always the desired speed
and not randomly faster or slower based on the angle of the ball.